Move `Profile` models / views to its own Django application
