+++ mangaki/mangaki/models.py
@@ -367,19 +367,6 @@ class Page(models.Model):
         return self.name
 
 
-class Profile(models.Model):
-    user = models.OneToOneField(User, on_delete=models.CASCADE)
-    is_shared = models.BooleanField(default=True)
-    nsfw_ok = models.BooleanField(default=False)
-    newsletter_ok = models.BooleanField(default=True)
-    reco_willsee_ok = models.BooleanField(default=False)
-    avatar_url = models.CharField(max_length=128, default='', blank=True, null=True)
-    mal_username = models.CharField(max_length=64, default='', blank=True, null=True)
-
-    def get_anime_count(self):
-        return Rating.objects.filter(user=self.user, choice__in=['like', 'neutral', 'dislike', 'favorite']).count()
-
-
 class Suggestion(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     work = models.ForeignKey(Work, on_delete=models.CASCADE)
