+++ mangaki/mangaki/tests/test_work.py
@@ -2,8 +2,8 @@ from django.contrib.auth.models import AnonymousUser
 from django.db import IntegrityError
 from django.test import TestCase
 
-from mangaki.factories import create_user_with_profile
 from mangaki.models import Category, Editor, Studio, Work
+from profiles.factories import create_user_with_profile
 
 
 class WorkTest(TestCase):
