+++ mangaki/mangaki/urls.py
@@ -16,6 +16,8 @@ urlpatterns = [
     url(r'^jsreverse/$', urls_js, name='js_reverse'),
 
     url(r'^$', views.index, name='home'),
+    url(r'^profiles/', include('profiles.urls', namespace='profiles')),
+    url(r'^u/(?P<username>.+)$', views.get_profile, name='profile'),
     url(r'^data/(?P<category>\w+)\.json$', views.get_works, name='get-work'),
     url(r'^data/reco/(?P<algo>\w+)/(?P<category>\w+)\.json$', views.get_reco_algo_list, name='get-reco-algo-list'),
     url(r'^data/reco_dpp/(?P<category>\w+)\.json$', views.get_reco_list_dpp, name='get-dpp-list'),
@@ -39,10 +41,6 @@ urlpatterns = [
     url(r'^artist/(?P<artist_id>\d+)/add/(?P<work_id>\d+)$', views.add_pairing, name='add-pairing'),
     url(r'^vote_dpp/(?P<work_id>\d+)$', views.dpp_work, name='vote-dpp'),
     url(r'^vote/(?P<work_id>\d+)$', views.rate_work, name='vote'),
-    url(r'^shared/$', views.update_shared, name='shared'),
-    url(r'^nsfw/$', views.update_nsfw, name='nsfw'),
-    url(r'^newsletter/$', views.update_newsletter, name='newsletter'),
-    url(r'^reco_willsee/$', views.update_reco_willsee, name='reco-willsee'),
     url(r'^mal/(?P<mal_username>.+)$', views.import_from_mal, name='import-mal'),
     url(r'^admin/', admin.site.urls),
     url(r'^about/$', views.about, name='about'),
