+++ mangaki/profiles/apps.py
@@ -1,0 +1,12 @@
+from django.apps import AppConfig
+
+
+class ProfilesConfig(AppConfig):
+    name = 'profiles'
+    verbose_name = 'Mangaki profiles'
+
+    def ready(self):
+        # Register receivers.
+        # noinspection PyUnresolvedReferences
+        import profiles.receivers
+
