+++ mangaki/profiles/receivers.py
@@ -2,7 +2,7 @@ from django.db.models.signals import post_save
 from django.dispatch import receiver
 from django.conf import settings
 
-from mangaki.models import Profile
+from profiles.models import Profile
 
 
 @receiver(post_save, sender=settings.AUTH_USER_MODEL)
